fixup! fixup! KTOR-4533 Add access to ApplicationConfig in plugins
