+++ ktor-server/ktor-server-core/jvm/test-resources/empty-config.yaml
@@ -1,0 +1,1 @@
+ktor: empty
