+++ ktor-server/ktor-server-core/jvm/test-resources/test-config.yaml
@@ -1,3 +1,7 @@
 
 myplugin:
     property: 42
+
+db:
+    myplugin:
+        property: 43
